pyproject.toml
[tool.poetry]
name = "select-things"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
pyyaml = "^6.0.1"
flask-sqlalchemy = "^3.1.1"
pymysql = "^1.1.1"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
